Removed nose from the package list - it just doesn't work that well.
nose was removed from the packages as it does not work cross-platform and
the command set is/appears to be pretty limited. Invoking the tests using
M-x compile gives much more flexibilty and allows interesting options such
as running coverage.

Also added command to remove whitespace when saving a python model file.