Freshened setup.
Reorganised the various setup files so they are more clearly linked
to customising parts of emacs itself or a major mode.

Heavily based on http://aaronbedra.com/emacs.d/ though the org-mode
setup was retained more or less in its entireity.

Everything loads with no problems however there may still be a
small amount of cruft since almost all of the code was 'borrowed'
and not written from scratch.